IteratedVariables class is now part of model instead of optimizer
